add springdoc to the project

+++ src/main/resources/db/migration/V20220519102825__init.sql
@@ -3,11 +3,11 @@ create table if not exists loan
     id                    uuid not null primary key,
     created_on            timestamp,
     updated_on            timestamp,
-    amount                numeric,
-    interest_rate         integer,
-    number_of_months      bigint,
+    amount                decimal,
+    interest_rate         float,
+    number_of_months      integer,
     payment_frequency     text,
-    total_interest_amount numeric
+    total_interest_amount decimal
 );
 alter index if exists loan_pkey rename to ind_l_id;
 
@@ -16,8 +16,8 @@ create table if not exists installment
     id              uuid not null primary key,
     created_on      timestamp,
     updated_on      timestamp,
-    amount          numeric,
-    number_of_month bigint,
+    amount          decimal,
+    number_of_month integer,
     loan_id         uuid not null,
     constraint fk_i_loan_id foreign key (loan_id) references loan (id)
 );
